requirements.txt
# ── LLM ───────────────────────────────────────────────────────────────────────
anthropic>=0.40.0
openai>=1.0.0

# ── Storage ────────────────────────────────────────────────────────────────────
sqlite-vec>=0.1.0
fastembed>=0.4.0

# ── Document loaders ──────────────────────────────────────────────────────────
pypdf>=4.0.0
docx2txt>=0.8
python-docx>=1.1.0

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0

# ── CLI ───────────────────────────────────────────────────────────────────────
prompt_toolkit>=3.0.0

# ── Telegram ──────────────────────────────────────────────────────────────────
python-telegram-bot>=20.0

# ── Research / data collection ────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
pytrends>=4.9.0
ddgs>=0.1.0
tavily-python>=0.3.0

# ── Utilities ─────────────────────────────────────────────────────────────────
dateparser>=1.2.0
filelock>=3.12.0
fpdf2>=2.7.0
croniter>=2.0.0
